refactor: ♻️ (web) コンポーネントのディレクトリ構造を修正した。  (#48)
- `index.ts`から再エクスポートするように変更し、内部のファイル命名を外部から気にする必要をなくした。
- 汎用性があるため`<Skeleton>`を`@/components`に移した。
- トップページ以外でも使われる装飾用であるため`<Aurora>`を`@/features/decoration/components`に移した。
