+++ apps/web/src/app/[statLabel]/page.tsx
@@ -2,8 +2,7 @@ import type { Metadata } from 'next';
 import type { ReactElement } from 'react';
 import { Suspense } from 'react';
 import { metadata as layoutMetadata, defaultTitle } from '../layout';
-import { PopulationChart } from '@/features/population/components/PopulationChart/PopulationChart';
-import { PopulationChartSkeleton } from '@/features/population/components/PopulationChart/PopulationChart.skeleton';
+import { PopulationChart, PopulationChartSkeleton } from '@/features/population/components/PopulationChart';
 import { fetchPrefectures } from '@/infra/resas/fetchPrefectures';
 import { getGraphPageTitleLocaleJa } from '@/libs/getGraphPageTitleLocale';
 import type { PrefCode } from '@/models/prefCode';
