+++ apps/web/src/components/Chart/index.ts
@@ -1,0 +1,4 @@
+export { Chart } from './Chart';
+export type { ChartProps } from './Chart';
+export { getUniqueLineColor } from './lineColors';
+export type { LineColor } from './lineColors';
