+++ apps/web/src/components/Image/index.ts
@@ -1,0 +1,2 @@
+export { Image } from './Image';
+export type { ImageProps, ImageSizeDimension } from './Image';
