+++ apps/web/src/components/Link/index.ts
@@ -1,0 +1,2 @@
+export { Link } from './Link';
+export type { LinkProps } from './Link';
