+++ apps/web/src/components/Select/index.ts
@@ -1,0 +1,16 @@
+export {
+  Select,
+  SelectGroup,
+  SelectValue,
+  SelectTrigger,
+  SelectIcon,
+  SelectViewport,
+  SelectPortal,
+  SelectContent,
+  SelectItemIndicator,
+  SelectItem,
+  SelectItemText,
+  SelectItemDescription,
+  SelectLabel,
+  SelectSeparator,
+} from './Select';
