+++ apps/web/src/features/decoration/components/Aurora/Aurora.story.tsx
