+++ apps/web/src/features/decoration/components/Aurora/Aurora.tsx
