+++ apps/web/src/features/decoration/components/Aurora/index.ts
@@ -1,0 +1,2 @@
+export { Aurora } from './Aurora';
+export type { AuroraProps } from './Aurora';
