+++ apps/web/src/features/navigation/components/Header/index.ts
@@ -1,0 +1,2 @@
+export { Header } from './Header';
+export type { HeaderProps } from './Header';
