+++ apps/web/src/features/navigation/components/PrefectureCheckbox/index.ts
@@ -1,0 +1,2 @@
+export { PrefectureCheckbox } from './PrefectureCheckbox';
+export type { PrefectureCheckboxProps } from './PrefectureCheckbox';
