+++ apps/web/src/features/navigation/components/PrefectureCheckboxFieldset/index.ts
@@ -1,0 +1,2 @@
+export { PrefectureCheckboxFieldset } from './PrefectureCheckboxFieldset';
+export type { PrefectureCheckboxFieldsetProps } from './PrefectureCheckboxFieldset';
