+++ apps/web/src/features/navigation/components/PrefectureStateBar/index.ts
@@ -1,0 +1,2 @@
+export { PrefectureStateBar, SELECTION_STATE_LABEL_ID } from './PrefectureStateBar';
+export type { PrefectureStateBarProps } from './PrefectureStateBar';
