+++ apps/web/src/features/navigation/components/ThemeSelect/index.ts
@@ -1,0 +1,2 @@
+export { ThemeSelect } from './ThemeSelect';
+export type { ThemeSelectProps } from './ThemeSelect';
