+++ apps/web/src/features/navigation/components/TopNavigation/index.ts
@@ -1,0 +1,2 @@
+export { TopNavigationLink, TopNavigationLinkFallback } from './TopNavigationLink';
+export type { TopNavigationLinkProps } from './TopNavigationLink';
