+++ apps/web/src/features/population/components/PopulationChart/index.ts
@@ -1,0 +1,3 @@
+export { PopulationChart } from './PopulationChart';
+export type { PopulationChartProps } from './PopulationChart';
+export { PopulationChartSkeleton } from './PopulationChart.skeleton';
